Create EECS3461_ExcelEntries.py
This is a simple Python script that was written to help my EECS 3461 group complete an assignment regarding Dark Pattern Analysis.
To be specific, a dataset we were given contained multiple duplicate entries. The point of this script is to analyze the given .xlsx file, and create a new page 
containing unique entries. The duplicate DE-ID tags were also highlighted, so as to avoid repeating a question, and keeping the inputs of the assignment consistent with each other across the board.